requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
geoalchemy2==0.15.2
pydantic==2.9.2
pydantic-settings==2.6.1
pyjwt==2.9.0
shapely==2.0.6
geopy==2.4.1
faker==30.8.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
email-validator==2.2.0
pytest
supabase==2.10.0